requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
sqlparse==0.5.1
